Add files via upload
################### C Language Assignment Solutions  ##############################

This repository contains my solutions to various assignments in C programming language. These solutions are aimed at helping others understand core concepts of C programming, including but not limited to:

- Variables and data types
- Control flow (if-else statements, loops)
- Functions and modular programming
- Arrays and strings
- Pointers and dynamic memory allocation
- File handling

Each assignment solution is organized into its own folder, with a clear and concise explanation provided along with the code. This repository is intended to serve as a resource for beginners learning C programming, offering them clear and commented solutions to common problems.